# spec:/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Roadmap Platform API
  version: 0.1.0
  description: |
    External REST API (BFF prefix `/api`). See `docs/05-api-contracts.md`.

    **Domain rules (P4):** Roadmap items reject initiative/roadmap pairs from different workspaces.
    Strategic themes with a `roadmapId` must target a roadmap in the same workspace. Initiative
    sponsors must belong to the initiative workspace. List endpoints support optional `?workspaceId=`
    for tenant-scoped reads.
servers:
  - url: http://localhost:4000/api
    description: Local API gateway
...
paths:
...
  /workspaces/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Workspaces]
      summary: Update workspace
    delete:
      tags: [Workspaces]
      summary: Delete empty workspace (cannot delete slug `default` or if data exists)
  /roadmaps:
    get:
      tags: [Roadmaps]
      summary: List roadmaps
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema: { type: string }
          description: Filter to roadmaps in this workspace.
    post:
      tags: [Roadmaps]
      summary: Create roadmap
  /roadmaps/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Roadmaps]
      summary: Get roadmap with items
    patch:
      tags: [Roadmaps]
      summary: Partial update roadmap
...
  /roadmaps/{id}/executive-summary:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Roadmaps]
      summary: Executive summary (themes, initiatives, phase health) for deck export
...
  /themes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Themes]
      summary: Get strategic theme with linked initiatives
    patch:
      tags: [Themes]
      summary: Update strategic theme
...